Updated with-superfine example so render is a type of side-effect rather than a message/update.
