+++ examples/with-superfine/package.json
@@ -11,7 +11,7 @@
   "license": "MIT",
   "dependencies": {
     "superfine": "^6.0.1",
-    "ferp": "^0.1.1"
+    "ferp": "^1.0.0"
   },
   "devDependencies": {
     "parcel-bundler": "^1.9.7"
